Adjustments to move quest_ip_tracking to Shard DB from World DB
